requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
websockets==12.0
python-multipart==0.0.9
pydantic-settings==2.2.1
httpx==0.24.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0 